Added some more implementatiton of Circuit model
